// .PRD/erp-refactor/v3/types/subcontracting-order.ts
import { z } from 'zod';

export const SubcontractingOrderSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  docstatus: z.number().default(0),
  title: z.string().optional().default('{supplier_name}'),
  naming_series: z.enum(['SC-ORD-.YYYY.-']),
  purchase_order: z.string(),
  supplier: z.string(),
  supplier_name: z.string(),
  supplier_warehouse: z.string(),
  supplier_currency: z.string().optional(),
  company: z.string(),
  transaction_date: z.string().default('Today'),
  schedule_date: z.string().optional(),
  amended_from: z.string().optional(),
  cost_center: z.string().optional(),
  project: z.string().optional(),
  set_warehouse: z.string().optional(),
  items: z.array(z.unknown()),
  total_qty: z.number().optional(),
  total: z.number().optional(),
  service_items: z.array(z.unknown()),
  set_reserve_warehouse: z.string().optional(),
  reserve_stock: z.boolean().optional().default(false),
  supplied_items: z.array(z.unknown()).optional(),
  supplier_address: z.string().optional(),
  address_display: z.string().optional(),
  contact_person: z.string().optional(),
  contact_display: z.string().optional(),
  contact_mobile: z.string().optional(),
  contact_email: z.string().optional(),
  shipping_address: z.string().optional(),
  shipping_address_display: z.string().optional(),
  billing_address: z.string().optional(),
  billing_address_display: z.string().optional(),
  distribute_additional_costs_based_on: z.enum(['Qty', 'Amount']).optional().default('Qty'),
  additional_costs: z.array(z.unknown()).optional(),
  total_additional_costs: z.number().optional(),
  status: z.enum(['Draft', 'Open', 'Partially Received', 'Completed', 'Material Transferred', 'Partial Material Transferred', 'Cancelled', 'Closed']).default('Draft'),
  per_received: z.number().optional(),
  select_print_heading: z.string().optional(),
  letter_head: z.string().optional(),
  production_plan: z.string().optional(),
});

export type SubcontractingOrder = z.infer<typeof SubcontractingOrderSchema>;

export const SubcontractingOrderInsertSchema = SubcontractingOrderSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type SubcontractingOrderInsert = z.infer<typeof SubcontractingOrderInsertSchema>;